Intermediate patch
This just gets pubspec.yaml correct, and the .html files for
verification and unit tests not hardcoding the path to dart.js. There
are no compilation errors, however, it still doesn't work. This check
in is necessary to solicit feedback from the community on the gap to
the next version.